# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Verity API
  description: |
    AI Content Authenticity Detection API.

    Verity provides multi-signal detection for AI-generated content across all media types,
    combined with cryptographic provenance verification.

    ## Authentication
    All endpoints require Bearer token authentication with an API key.

    ## Rate Limits
    Rate limits are enforced per API key based on the key's tier.
  version: 1.0.0
  contact:
    email: api@verity.ai

servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.verity.ai
    description: Production

security:
  - bearerAuth: []

tags:
  - name: detection
    description: AI content detection endpoints
  - name: jobs
    description: Async job management
  - name: keys
    description: API key management
  - name: compat
    description: AIorNot-compatible endpoints
  - name: system
    description: System health and info

paths:
  /api/v1/detect/image:
    post:
      tags: [detection]
      summary: Detect AI in image
      operationId: detectImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                only:
                  type: string
                  description: Comma-separated list of detectors to run
                excluding:
                  type: string
                  description: Comma-separated list of detectors to exclude
                external_id:
                  type: string
                  description: Your tracking ID
      responses:
        '200':
          description: Detection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionResponse'

  /api/v1/detect/text:
    post:
      tags: [detection]
      summary: Detect AI-generated text
      operationId: detectText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                include_annotations:
                  type: boolean
                  default: false
                external_id:
                  type: string
      responses:
        '200':
          description: Detection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionResponse'

  /api/v1/jobs:
    post:
      tags: [jobs]
      summary: Create async detection job
      operationId: createJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                webhook_url:
                  type: string
                  format: uri
      responses:
        '202':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreated'

  /api/v1/jobs/{id}:
    get:
      tags: [jobs]
      summary: Get job status and results
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'

  /v2/image/sync:
    post:
      tags: [compat]
      summary: AIorNot-compatible image detection
      operationId: compatImageSync
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Detection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompatImageResponse'

  /v1/system/live:
    get:
      tags: [system]
      summary: System live check
      operationId: systemLive
      security: []
      responses:
        '200':
          description: System is live
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_live:
                    type: boolean

  /healthz:
    get:
      tags: [system]
      summary: Health check
      security: []
      responses:
        '200':
          description: Service is healthy

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    DetectionResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        media_type:
          type: string
          enum: [image, video, audio, text]
        verdict:
          type: string
          enum: [ai, human, uncertain, unavailable]
        confidence:
          type: number
        signals:
          type: array
          items:
            $ref: '#/components/schemas/SignalResult'

    SignalResult:
      type: object
      properties:
        name:
          type: string
        verdict:
          type: string
        confidence:
          type: number
        model_version:
          type: string

    JobCreated:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string

    JobResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        result:
          $ref: '#/components/schemas/DetectionResponse'

    CompatImageResponse:
      type: object
      properties:
        id:
          type: string
        report:
          type: object
          properties:
            verdict:
              type: string
            ai:
              type: object
              properties:
                is_detected:
                  type: boolean
                confidence:
                  type: number